Add automatic flashing of Home Assistant SkyConnect/fix compatibility with ZHA config flow (#86)
* Add automatic flashing of Home Assistant SkyConnect

Implement automatic flashing of Home Assistant SkyConnect. Currently
this is behind a flag which defaults to off.

* Use oneshot service for flashing

Using oneshot service for flashing is much more suitable. This also
fixes disabled flashing.

* Do not restart zigbeed on connection close

Instead of restarting on connection close, just listen on the same port
again. This matches the behavior of a serial port more, and allows ZHA
to reconnect immeadily on disconnect (required during discovery).

* Update CHANGELOG.md

* Get mDNSResponder from alternative source

* Pre-install some Python package/fix build for armv7 by adding python3-dev

* Add Home Assistant Yellow support

* Add documentation hints

* Mention Yellow in the changelog