matter_server_beta: remove beta add-on
Maintaining Matter Server beta as a separate add-on has the problem that
the Matter fabric data is separated from the official/non-beta add-on.
Furthermore the Matter integration installs and starts the official
add-on which causes furhter complications. Adjusting the URL to use the
beta add-on is also non-trivial.

For these reasons, the offical Matter Server add-on received a beta flag
a while ago (see https://github.com/home-assistant/addons/pull/3282).
Drop the pure beta version as it is no longer required.